+++ mopidy/core/__init__.py
@@ -5,6 +5,7 @@ from .actor import Core
 from .history import HistoryController
 from .library import LibraryController
 from .listener import CoreListener
+from .mixer import MixerController
 from .playback import PlaybackController, PlaybackState
 from .playlists import PlaylistsController
 from .tracklist import TracklistController